Moved operators testing into its own file(s) and improved parse_show_table_col_attrs() utils function to work with quoted column names (with spaces).
